Vast improvements to the Player Id lookup system
Something that I spent a bit on making. This is a proposal to replace the old Player Id lookup with one with far more functionality and player Ids with greatly improved searching abilities. This allows the user to use multiple parameters to search, as well as run faster as the player ID map downloads locally.